[mural] extend timeout
Summary:
I was running into errors where we'd timeout waiting for the response
from the server process and then kill it, which just breaks
everything.  Instead extend the timeout to something longer (.5s) and
warn when its hit.

Also add a debugging flag to optionally echo the server response to
the *muralserver* buffer.

Test Plan: Used it for a while, no more timeouts but still feels responsive.