[mural] Store more metadata from the tags file
Summary:
In addition to symbols, store the filename and row for each tag.  This
will be used in later diffs to replace find-tag and to internally
reindex as the files change.

Additionally, avoid unnesary copies in find_fuzzy_match() by passing a
pointer to the TagInfo (previously std::string).  This takes the 'a'
query to 67ms.  This moves the sort+limit inside find_fuzzy_matches
to make sure none of the pointers escape.

Test Plan: Verified 'a' query was down to 67ms.