[mural] Fix rotation
Summary:
ido-next-match should rotate the results so the next result becomes
first.  This wasn't working because the ido-dynamic-match override was
resetting the result set on every call.  To fix this, only run the
query if the query has changed.

Test Plan: Was able to use C-s and [right] to rotate results.