[mural] Perf optimizations
Summary:
Search was still a little laggy when typing since each keypress is
~130ms.  The main change is to store a bool for whether each tag is/is
not a method at read time, and only calculate whether the query is a
method query once.  We're still iterating over the entire result set
but this about 4x faster.  A further optimization is to actually split
into method and non-method tag sets, but ATM that's unnessary.

The main perf issue now is the call to find-tag in emacs.

Test Plan: Ran from the command line, verified timings were ~4x faster.