[mural] bypass find-tag
Summary:
mural.el was using find-tag to actually open the tag we selected.
This often incurs an expensive re-reads of the TAGS file and is far
slower than the actual search at this point.  Bypass the tags file
entirely by returning the file and line# from the server and open the
file directly.

Things will get a little complicated when we eventually want to
support multiple repos at the same time since this only expects one
tags-file-name, but we'll deal with that later.

Test Plan: C-o CDT::cacheKeys, it opened the right thing.