[mural] Multiple repo support and doc updates
Summary:
Adds support for multiple repositories/tag files.  Tag files are now
specified by

  (mural-add-tagfile "~/www/TAGS")
  (mural-add-tagfile "~/local/hphp/src/TAGS")

The tagfile/repo to choose is based on the location of the currently
open file (e.g files in hphp/src search that tags file, files in ~/www
use that one).

Test Plan:
C-o in an directory without a tags file returns error

Opening a file in hphp/src or www/ use the associated tags file.

Killing the muralserver buffer/proces does not lead to a leak in the
tagfile-to-process alist.