[mural] case-insensitive search
Summary:
Search is more useful if its case-insensitive, at least when the
entire search string is all lower case.

This is a bit of a perf hit (35ms -> 50ms), but within the acceptable
range still.  I'm making a memory/speed tradeoff since converting to
lower-case at query time takes this to 177ms and that is noticable.

Test Plan: coredatatype -> finds CoreDataType