[mural] Qualify duplicate tag results
Summary:
A query like `URI` returns several duplicate results for functions,
modules, and classes in both the JS and PHP codebases.  This is not
only confusing, it's impossible to cycle through the results to select
the next one when this happens.  This change qualifies the displayed
name of a tag to include the filename, or full path if necessary.

Test Plan:
A search for URI returns several unique results in both JS and php,
I can C-s through the list, and selecting one takes me to the expected place.