[mural] Fall back to case-insensitive match if no results found
Summary:
Sometimes I typo my query by adding an extra capital letter, this does
the right thing in that case.  A more general solution is an edit
distance query but I'm not going to tackle that until the search is
run in parallel.

Test Plan:
Queried for a string with and without a misplaced capital letter